backend/requirements.txt
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# pydantic==2.5.0
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.9
# python-dotenv==1.0.0
# langchain==0.1.0
# langchain-openai==0.1.0
# langchain-community==0.0.20
# openai>=1.6.1
# python-multipart==0.0.6
# cors==1.0.1
# fastapi-cors==0.0.6
# python-jose[cryptography]==3.3.0
# email-validator==2.1.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai>=1.6.1
python-multipart==0.0.6
cors==1.0.1
fastapi-cors==0.0.6
python-jose[cryptography]==3.3.0
email-validator==2.1.0